# src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: API RESTful - DoaPix
  version: 1.0.0
  description: >-
    Este projeto consiste na construção de uma **API RESTful usando Node.js e
    TypeScript**, com integração ao banco de dados **PostgreSQL**, para criar um
    sistema que permita aos usuários gerenciar **doações solidárias online**. As
    doações serão organizadas em categorias e também terão um histórico para
    acompanhamento de doações realizadas. Através de uma interface amigável, os
    usuários poderão **criar, visualizar, editar e excluir doações**, além de
    acompanhar **doações previamente realizadas**.


    A API fornecerá aos usuários todas as funcionalidades necessárias para
    realizar operações básicas em suas doações, como criar, visualizar, editar e
    excluir. Além disso, ela será projetada para garantir a segurança e a
    privacidade dos dados do usuário, implementando práticas adequadas de
    autenticação e autorização.


    O uso do **Node.js e TypeScript** permitirá uma **fácil implementação dos
    endpoints da API**, bem como a manutenção do código com tipagem estática,
    tornando o processo de desenvolvimento mais seguro e organizado.
    Utilizaremos o **PostgreSQL** como nosso banco de dados para **armazenar as
    doações e as informações relacionadas a categorias e histórico de
    transações**.


    Para a **documentação da API**, faremos uso do **Swagger UI** através do
...
    entenderem e interagirem com os endpoints da API.


...


...
  license:
...
components:
  securitySchemes:
    jwt:
      type: apiKey
      in: header
      name: Authorization
      description: Para autenticar, coloque o token JWT no formato **Bearer _token_**.
  schemas:
    DonationList:
      type: array
      items:
        $ref: '#/components/schemas/Donation'
    DonationDetails:
      type: object
      required:
        - id
        - name
        - url_image
        - description
        - state
        - category
        - goal
        - amount_raised
        - deadline
        - created_at
        - user_id
        - transactions
        - comments
      properties:
        id:
          type: string
          description: Identificador único da campanha de doação
          example: fa763c0d-21ad-4151-bd14-00962d3c3a74
        name:
          type: string
          description: Nome da campanha de doação
          example: Doe Amor e Esperança
        url_image:
          type: string
          description: URL da imagem da campanha de doação (formato base64)
          example: >-
            data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wCEAAoHCBURDw8REREPDw8PDxEPEA8PEREPDw8PGBQZGRgUGBgcIS4lHB4sHxkYJjgnLTU0NjU1GiQ7QDszQC40Nz8BDAwMEA8QHBISGDQhJSExNDQxMTE0NDExNDQ1NDE0NDQ0NDQ0NDQxNDE0MTE0NDQ0NDQ0NDQ0MTQxNDE0NDQ3NP/AABEIAJwBRAMBIgACEQEDEQH/xAAbAAACAgMBAAAAAAAAAAAAAAABAgADBQYHBP/EAEsQAAIBAwIDBAYGBQkECwAAAAECAAMEEQUSBiExE0FRYQcUIjJxgUJSYnKRshcjM6GxFVNUdIKSlKLT
        description:
          type: string
          description: Descrição da campanha de doação
          example: >-
            Buscamos arrecadar fundos e itens essenciais para apoiar famílias
            carentes. As doações podem ser feitas pessoalmente ou por
            transferência bancária. Sua contribuição pode fazer a diferença!
        state:
          type: string
          description: >-
            Sigla dos estados do Brasil (Unidades federativas do Brasil), onde
            se localiza essa campanha de doação
          example: PR
        category:
          type: string
          description: Categoria que se encaixa essa campanha de doação
          example: Projetos Sociais / Voluntariado
        goal:
          type: string
          description: Meta de arrecadação monetária da campanha de doação
          example: $500.00
        amount_raised:
          type: string
          description: Valor monetária já arrecadado na campanha de doação
          example: $100.00
        deadline:
          type: string
          description: Data limite da campanha de doação, sendo fechada após esse período
          example: '2024-04-24T19:34:02.090Z'
        created_at:
          type: string
          description: Data de criação da campanha de doação
          example: '2024-03-24T19:34:02.090Z'
        user_id:
          type: string
          description: Usuário que criou essa campanha de doação
          example: f1cd0ab3-4f34-4a53-8b2c-594dd917d6ca
        transactions_count:
          type: integer
          description: Quantidade de doações pagas nessa campanha
          example: 15
        comments:
          type: array
          description: Lista de comentários solidários dos doadores
          items:
            type: object
            properties:
              alias:
                type: string
                description: Apelido do doador (opcional)
                example: João Teixeira
              message:
                type: string
                description: Mensagem solidária do doador
                example: 🤝 Fiz uma contribuição. Espero que ajude a fazer a diferença!
    Donation:
      type: object
      required:
        - id
        - name
        - url_image
        - description
        - state
        - category
        - goal
        - amount_raised
        - deadline
        - created_at
        - user_id
      properties:
        id:
          type: string
          description: Identificador único da campanha de doação
          example: fa763c0d-21ad-4151-bd14-00962d3c3a74
        name:
          type: string
          description: Nome da campanha de doação
          example: Doe Amor e Esperança
        url_image:
          type: string
          description: URL da imagem da campanha de doação (formato base64)
          example: >-
            data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wCEAAoHCBURDw8REREPDw8PDxEPEA8PEREPDw8PGBQZGRgUGBgcIS4lHB4sHxkYJjgnLTU0NjU1GiQ7QDszQC40Nz8BDAwMEA8QHBISGDQhJSExNDQxMTE0NDExNDQ1NDE0NDQ0NDQ0NDQxNDE0MTE0NDQ0NDQ0NDQ0MTQxNDE0NDQ3NP/AABEIAJwBRAMBIgACEQEDEQH/xAAbAAACAgMBAAAAAAAAAAAAAAABAgADBQYHBP/EAEsQAAIBAwIDBAYGBQkECwAAAAECAAMEEQUSBiExE0FRYQcUIjJxgUJSYnKRshcjM6GxFVNUdIKSlKLT
        description:
          type: string
          description: Descrição da campanha de doação
          example: >-
            Buscamos arrecadar fundos e itens essenciais para apoiar famílias
            carentes. As doações podem ser feitas pessoalmente ou por
            transferência bancária. Sua contribuição pode fazer a diferença!
        state:
          type: string
          description: >-
            Sigla dos estados do Brasil (Unidades federativas do Brasil), onde
            se localiza essa campanha de doação
          example: PR
        category:
          type: string
          description: Categoria que se encaixa essa campanha de doação
          example: Projetos Sociais / Voluntariado
        goal:
          type: string
          description: Meta de arrecadação monetária da campanha de doação
          example: $500.00
        amount_raised:
          type: string
          description: Valor monetária já arrecadado na campanha de doação
          example: $100.00
        deadline:
          type: string
          description: Data limite da campanha de doação, sendo fechada após esse período
          example: '2024-04-24T19:34:02.090Z'
        created_at:
          type: string
          description: Data de criação da campanha de doação
          example: '2024-03-24T19:34:02.090Z'
        user_id:
          type: string
          description: Usuário que criou essa campanha de doação
          example: f1cd0ab3-4f34-4a53-8b2c-594dd917d6ca
    DonationUpdate:
      type: object
      required:
        - name
        - url_image
        - description
        - state
        - category
        - goal
        - deadline
      properties:
        name:
          type: string
          description: Nome da campanha de doação
          example: Doe Amor e Esperança
        url_image:
          type: string
          description: URL da imagem da campanha de doação (formato base64)
          example: >-
            data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wCEAAoHCBURDw8REREPDw8PDxEPEA8PEREPDw8PGBQZGRgUGBgcIS4lHB4sHxkYJjgnLTU0NjU1GiQ7QDszQC40Nz8BDAwMEA8QHBISGDQhJSExNDQxMTE0NDExNDQ1NDE0NDQ0NDQ0NDQxNDE0MTE0NDQ0NDQ0NDQ0MTQxNDE0NDQ3NP/AABEIAJwBRAMBIgACEQEDEQH/xAAbAAACAgMBAAAAAAAAAAAAAAABAgADBQYHBP/EAEsQAAIBAwIDBAYGBQkECwAAAAECAAMEEQUSBiExE0FRYQcUIjJxgUJSYnKRshcjM6GxFVNUdIKSlKLT
        description:
          type: string
          description: Descrição da campanha de doação
          example: >-
            Buscamos arrecadar fundos e itens essenciais para apoiar famílias
            carentes. As doações podem ser feitas pessoalmente ou por
            transferência bancária. Sua contribuição pode fazer a diferença!
        state:
          type: string
          description: >-
            Sigla dos estados do Brasil (Unidades federativas do Brasil), onde
            se localiza essa campanha de doação
          example: PR
        category:
          type: string
          description: Categoria que se encaixa essa campanha de doação
          example: Projetos Sociais / Voluntariado
        goal:
          type: string
          description: Meta de arrecadação monetária da campanha de doação
          example: $500.00
        deadline:
          type: string
          description: Data limite da campanha de doação, sendo fechada após esse período
          example: '2024-04-24T19:34:02.090Z'
    DonationInsert:
      type: object
      required:
        - name
        - url_image
        - description
        - state
        - category
        - goal
        - deadline
      properties:
        name:
          type: string
          description: Nome da campanha de doação
          example: Doe Amor e Esperança
        url_image:
          type: string
          description: URL da imagem da campanha de doação (formato base64)
          example: >-
            data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wCEAAoHCBURDw8REREPDw8PDxEPEA8PEREPDw8PGBQZGRgUGBgcIS4lHB4sHxkYJjgnLTU0NjU1GiQ7QDszQC40Nz8BDAwMEA8QHBISGDQhJSExNDQxMTE0NDExNDQ1NDE0NDQ0NDQ0NDQxNDE0MTE0NDQ0NDQ0NDQ0MTQxNDE0NDQ3NP/AABEIAJwBRAMBIgACEQEDEQH/xAAbAAACAgMBAAAAAAAAAAAAAAABAgADBQYHBP/EAEsQAAIBAwIDBAYGBQkECwAAAAECAAMEEQUSBiExE0FRYQcUIjJxgUJSYnKRshcjM6GxFVNUdIKSlKLT
        description:
          type: string
          description: Descrição da campanha de doação
          example: >-
            Buscamos arrecadar fundos e itens essenciais para apoiar famílias
            carentes. As doações podem ser feitas pessoalmente ou por
            transferência bancária. Sua contribuição pode fazer a diferença!
        state:
          type: string
          description: >-
            Sigla dos estados do Brasil (Unidades federativas do Brasil), onde
            se localiza essa campanha de doação
          example: PR
        category:
          type: string
          description: Categoria que se encaixa essa campanha de doação
          example: Projetos Sociais / Voluntariado
        goal:
          type: string
          description: Meta de arrecadação monetária da campanha de doação
          example: $500.00
        deadline:
          type: string
          description: Data limite da campanha de doação, sendo fechada após esse período
          example: '2024-04-24T19:34:02.090Z'
    DeleteSuccess:
      type: object
      properties:
        msg:
          type: string
          description: Mensagem de sucesso ao excluir
          example: Excluído com sucesso
...
    BadRequest:
      type: object
      properties:
        msg:
          type: string
          description: >-
            Requisição inválida. Alguns parâmetros podem estar faltando ou serem
            inválidos
          example: Alguns parâmetros podem estar faltando ou serem inválidos
    Unauthorized:
      type: object
      properties:
        msg:
          type: string
          description: Não autorizado. O token JWT não foi fornecido ou é inválido
          example: O token JWT não foi fornecido ou é inválido
    Forbidden:
      type: object
      properties:
        msg:
          type: string
          description: >-
            Proibido. O usuário não tem permissão para acessar o recurso
            solicitado
          example: Não tem permissão para acessar o recurso solicitado
    NotFound:
      type: object
      properties:
        msg:
          type: string
          description: Nenhum dado encontrado. O recurso solicitado não existe
          example: Nenhum dado encontrado
    InternalServerError:
      type: object
      properties:
        msg:
          type: string
          description: Erro interno ao manipular dados no serviço
          example: Erro interno do servidor
    TransactionList:
      type: array
      items:
        $ref: '#/components/schemas/Transaction'
    Transaction:
      type: object
      required:
        - id
        - txid
        - location
        - qr_code
        - pix_copy_paste
        - amount
        - alias
        - email
        - message
        - status
        - created_at
        - donation_id
      properties:
        id:
          type: string
          description: Identificador único da transação
          example: f1cd0ab3-4f34-4a53-8b2c-594dd917d6ca
        txid:
          type: string
          description: >-
            Identificador único atribuído uma transação feita com uma
            criptomoeda na rede blockchain
          example: 7978c0c97ea847e78e8849634473c1f1
        location:
          type: string
          description: >-
            Identificador único associado a uma cobrança imediata no sistema de
            pagamentos instantâneos Pix, responsável por gerar o QR Code
            correspondente à cobrança
          example: pix.example.com/qr/v2/9d36b84fc70b478fb95c12729b90ca25
        qr_code:
          type: string
          description: Imagem base64 do QR Code para realizar o pagamento
          example: >-
            data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmc
            vMjAwMC9zdmciIHZpZXdCb3g9IjAgMCA0NSA0NSIgc2hhcGUtcmVuZGVyaW5nPSJjcmlzcEVkZ2VzIj48cGF0aCBmaWxsPSIjZmZmZmZmIiBkPSJNMCAwaDQ1djQ1SD
        pix_copy_paste:
          type: string
          description: >-
            String permite realizar pagamentos sem a necessidade de escanear um
            QR Code
          example: >-
            00020101021226830014BR.GOV.BCB.PIX2561qrcodespix.sejaefi.com.br/v2/41e0badf811a4ce6ad8a80b306821fce5204000053000065802BR5905EFISA6008SAOPAULO60070503***61040000
        amount:
          type: string
          description: Valor da transação
          example: $2,000.00
        alias:
          type: string
          description: Apelido do doador (opcional)
          example: João Teixeira
        email:
          type: string
          description: Email do doador
          example: email@example.com
        message:
          type: string
          description: Mensagem solidária do doador
          example: 🤝 Fiz uma contribuição. Espero que ajude a fazer a diferença!
        status:
          type: string
          description: Status da transação
          example: Pago
        created_at:
          type: string
          description: Data de criação da transação
          example: '2024-03-24T19:34:02.090Z'
        donation_id:
          type: string
          description: Identificador único da transação da campanha de doação
          example: fa763c0d-21ad-4151-bd14-00962d3c3a74
...
paths:
  /donations:
    get:
      summary: Vizualiza todas campanhas de doação
      description: Retorna as informações de todas campanhas de doação
      tags:
        - Donation
      operationId: get_all_donation
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationList'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    post:
      summary: Cria uma nova campanha de doação
      description: >-
        Cria uma nova campanha de doação com base nos dados fornecidos no corpo
        da requisição
      tags:
        - Donation
      operationId: create_donation
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DonationInsert'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donation'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /donations/{donation_id}:
    parameters:
      - name: donation_id
        in: path
        required: true
        description: ID da campanha de doação que será deletada
        schema:
          type: string
    get:
      summary: Visualiza campanha de doação pelo id
      description: Retorna as informações da campanha de doação pelo id
      tags:
        - Donation
      operationId: get_donation_by_id
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonationDetails'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    put:
      summary: Atualiza campanha de doação
      description: >-
        Atualiza as informações da campanha de doação desejada, desde que o
        usuário que está realizando a edição seja o mesmo que a criou
      tags:
        - Donation
      operationId: update_donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DonationUpdate'
      security:
        - jwt: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donation'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      summary: Remove campanha de doação
      description: >-
        Apaga campanha de doação, desde que o usuário que está realizando a
        exclusão seja o mesmo que a criou
      tags:
        - Donation
      operationId: delete_donation
      security:
        - jwt: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSuccess'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /transactions/donation/{donation_id}:
    parameters:
      - name: donation_id
        in: path
        required: true
        description: ID da campanha de doação que será obtido as transações
        schema:
          type: string
    get:
      summary: Vizualiza todas transações solidárias de uma campanha de doação
      description: >-
        Retorna as informações de todas transações solidárias de uma campanha de
        doação
      tags:
        - Transaction
      operationId: get_all_transaction_by_donation
      security:
        - jwt: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionList'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...